refactor: modified Tasklist page in components
